+++ CHANGES.md
@@ -30,6 +30,7 @@
   - ➕ `slice` => `jsSlice` / `js_slice`
   - ➕ `substring`
   - ➕ `toLocaleLowerCase` / `to_locale_lower_case`
+  - ➕ `toLocaleUpperCase` / `to_locale_upper_case`
 
 ## 0.0.1 - 2023-04-28